Incremental change to make the processing of SVN dump files more fine grained (down from SvnRevision to SvnNode). * Move TerminatingValidator out of test source into main source, as it might be useful.
