+++ src/main/java/com/github/cstroe/svndumpgui/internal/transform/NoopSvnDumpMutator.java
@@ -2,6 +2,7 @@ package com.github.cstroe.svndumpgui.internal.transform;
 
 import com.github.cstroe.svndumpgui.api.SvnDumpMutator;
 import com.github.cstroe.svndumpgui.api.SvnDumpPreamble;
+import com.github.cstroe.svndumpgui.api.SvnNode;
 import com.github.cstroe.svndumpgui.api.SvnRevision;
 
 public class NoopSvnDumpMutator implements SvnDumpMutator {
@@ -10,6 +11,9 @@ public class NoopSvnDumpMutator implements SvnDumpMutator {
 
     @Override
     public void consume(SvnRevision revision) {}
+
+    @Override
+    public void consume(SvnNode node) {}
 
     @Override
     public void finish() {}
