+++ src/main/java/com/github/cstroe/svndumpgui/internal/validate/AbstractSvnDumpValidator.java
@@ -2,6 +2,7 @@ package com.github.cstroe.svndumpgui.internal.validate;
 
 import com.github.cstroe.svndumpgui.api.SvnDumpPreamble;
 import com.github.cstroe.svndumpgui.api.SvnDumpValidator;
+import com.github.cstroe.svndumpgui.api.SvnNode;
 import com.github.cstroe.svndumpgui.api.SvnRevision;
 
 public abstract class AbstractSvnDumpValidator implements SvnDumpValidator {
@@ -10,6 +11,9 @@ public abstract class AbstractSvnDumpValidator implements SvnDumpValidator {
 
     @Override
     public void consume(SvnRevision revision) {}
+
+    @Override
+    public void consume(SvnNode node) {}
 
     @Override
     public void finish() {}
