+++ src/main/java/com/github/cstroe/svndumpgui/internal/validate/TerminatingValidator.java
@@ -1,11 +1,11 @@
-package com.github.cstroe.svndumpgui.internal.utility;
+package com.github.cstroe.svndumpgui.internal.validate;
 
 import com.github.cstroe.svndumpgui.api.SvnDumpError;
 import com.github.cstroe.svndumpgui.api.SvnDumpPreamble;
 import com.github.cstroe.svndumpgui.api.SvnDumpValidator;
 import com.github.cstroe.svndumpgui.api.SvnRevision;
 
-public class TerminatingValidator implements SvnDumpValidator {
+public class TerminatingValidator extends AbstractSvnDumpValidator {
 
     private final SvnDumpValidator wrappedValidator;
