+++ src/main/java/com/github/cstroe/svndumpgui/internal/writer/AbstractSvnDumpWriter.java
@@ -2,6 +2,7 @@ package com.github.cstroe.svndumpgui.internal.writer;
 
 import com.github.cstroe.svndumpgui.api.SvnDumpPreamble;
 import com.github.cstroe.svndumpgui.api.SvnDumpWriter;
+import com.github.cstroe.svndumpgui.api.SvnNode;
 
 import java.io.OutputStream;
 import java.io.PrintStream;
@@ -27,6 +28,9 @@ public abstract class AbstractSvnDumpWriter implements SvnDumpWriter {
 
     @Override
     public void consume(SvnDumpPreamble preamble) {}
+
+    @Override
+    public void consume(SvnNode node) {}
 
     @Override
     public void finish() {}
